Update shortchut in the creature page to use the creatures dex order
